refactor Manifest class and some optimizations
This creates a more generic Manifest type and shows
how one can use this generic type to create a specialized
type, such as the one used for English ASR.

Secondly, the names of the variables was not consistent
with what the data actually represented.  Particularly,
the encoded tokens have changed names from `transcript` to
`tokens` to make it more clear that this is an encoded
representation of the text and not a text transcript.

Lastly, the code for the collate functions seemed relatively
inefficient and made a lot of assumptions regarding data types
and device.  There are changes which should work regardless
of the data types of the items in the batch.

Signed-off-by: David Pollack <david@i2x.ai>
Signed-off-by: Tomasz Kornuta <tkornuta@nvidia.com>