Add gpus-per-node parameter to GPT-3
The new --gpus-per-node parameter for Slurm clusters was missing from
the conversion and evaluation scripts for GPT-3, resulting in the
containers launching without GPUs and failing for those phases. This
allows users to specify the total number of GPUs to be included in the
node which is the most flexible with different Slurm-based clusters.

Signed-Off-By: Robert Clark <roclark@nvidia.com>