Update default GPU specification
The previous default value of gpus_per_task=1 was not as compatible
between different Slurm clusters depending on how the cgroups were
configured. If not using an identical cgroup configuration as our
test clusters, the expected number of GPUs might not be present inside
running containers. gpus_per_node=8 is more flexible with different
cgroup settings while offering the same functionality and performance
as gpus_per_task=1 and should be used instead to offer the highest
level of compatibility.

Signed-Off-By: Robert Clark <roclark@nvidia.com>