Add warning when model is not in eval mode during generation
Systematically calling `mode.eval()` does not seem like a good idea, as
it might have side effects leading to unexpected behavior. It would be
better to raise an exception if one attempts to generate while in
training mode, but this may break existing code => sticking to a warning
for now.

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>