Bugfix and optimization in `end_of_generation_condition()` (#7267)
* Bugfix and optimization in `end_of_generation_condition()`

1. Bugfix

The previous implementation did not verify that `end_string` was encoded
into a single token, which could trigger the end of generation earlier
than intended (see discussion in #7187)

2. Optimization

The previous implementation was scaling linearly with the batch size and
quadratically with the length of the generated sequence, which could
lead to a significant overhead in some situations.

The new implementation is much more efficient in "normal" situations
(where the end of generation is identified by a set of unique tokens),
and raises a warning when it needs to fallback to the inefficient string
matching case.

Note that it does not behave exactly the same as before, because we skip
the string comparison when the end strings all have unique tokens
associated to them. For instance, in the previous implementation, if the
model had generated the string
        "Some string.<|endoftext|>"
(where "<|endoftext|>" would really be generated as a string, and not as
a single token), then the previous implementation would have considered
it to be the end of generation (assuming `end_strings` has length > 1),
while the new one would not. The previous behavior was likely a bug
though, since we expect models to generate the special tokens associated
to end strings when they exist (for instance, the standard case
`end_strings=["<|endoftext|>"]` has always been handled by just
comparing the last token to `eod_id`).

Fixes #7187

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>

* Add warning when model is not in eval mode during generation

Systematically calling `mode.eval()` does not seem like a good idea, as
it might have side effects leading to unexpected behavior. It would be
better to raise an exception if one attempts to generate while in
training mode, but this may break existing code => sticking to a warning
for now.

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>

* Use "<extra_id_1>" as prefix string

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>

* Add TODO for potential failure mode of the string match mechanism

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>

---------

Signed-off-by: Olivier Delalleau <507137+odelalleau@users.noreply.github.com>