Update the TopKSequenceGenerator to support Top-P sampling
Also known as nucleus sampling. Put it in with the Top-K sampler, as
I felt like this kept the code cleaner – at the cost of a single if
statement in the decoding loop. Could move it to its own class as well,
but since a lot of people use Top-K and Top-P together, we'd probably
want that new class to support both anyway.