Support S3 data loading
This commit introduces the S3IndexedDataset, which supports
loading a dataset stored in S3 in the same format as the
MMapIndexedDataset. In particular, the .idx file is downloaded
to a local directory at initialization so that we can memory map
it and the .bin file is streamed into memory block-by-block.

In order for S3 data loading to be performant, we introduce block
shuffling to GPTDataset, where we divide samples into blocks,
shuffle the blocks and then shuffle within the blocks. In this
way, we introduce a minimal amount of correlation into the
shuffle, which can be exploited to download larger,
continguous chunks of data from S3. We set the default block
size for shuffling to 1, which recovers the shuffling behavior
before this commit.