Dgalvez/cuda graphs greedy rnnt inference squash (#8191)
* Speed up RNN-T greedy decoding with cuda graphs

This uses CUDA 12.3's conditional node support.

Initialize cuda tensors lazily on first call of __call__ instead of __init__.

We don't know what device is going to be used at construction time,
and we can't rely on torch.nn.Module.to() to work here. See here:
https://github.com/NVIDIA/NeMo/issues/8436

This fixes an error "Expected all tensors to be on the same device,
but found at least two devices" that happens when you call to() on your
torch.nn.Module after constructing it.

https://github.com/NVIDIA/NeMo/pull/8191#discussion_r1491227311

Signed-off-by: Daniel Galvez <dgalvez@nvidia.com>
Signed-off-by: Pablo Garay <pagaray@nvidia.com>