test_step(): handle invalid audio
When the model generates an output that is very short (less than 2 ASR frames) the ASR and SSIM calculations will error out. We detect the error and invalidate the entire batch, setting WER/CER to 100% and SSIM to 0.0. The transcription is set to "<INVALID">.

Note the metrics still written out to the `.metrics` files; they need to be ignored by any subsequent statistics calculations.