Safely import nvidia-resiliency-ext package
The nvidia-resiliency-ext package is not compatible with Windows/MacOS
currently and cannot be installed directly on those platforms. Given the
number of nested imports in the nemo.collections.llm API, it will try and
automatically import the resliency package even if not used. Safely
importing the package allows all modules to be imported without blocking
on the resiliency package for incompatible platforms. This opens up
greater support on more platforms.

If the nvidia-resiliency-ext package is needed when these lines are called,
it will throw an import error only when directly called.

Signed-Off-By: Robert Clark <roclark@nvidia.com>