Safely import modelopt algorithms
The default modelopt model transforms rely on the modelopt package being
installed which is not available on all platforms. This module gets
auto-imported when importing the nemo.collections.llm module and will
fail on platforms that don't support modelopt, such as Windows and MacOS,
even if the package isn't used.

Safely importing the package should allow the toolkit to be runnable on
both Windows and MacOS in a minimal state, which increases its
functionality, especially while paired with NeMo-Run.

Signed-Off-By: Robert Clark <roclark@nvidia.com>