Don't allow EOS until 4 frames have been generated
The number of frames is configuratble via a parameter to infer_batch().

This is a workaround to the observation that when CFG is on
we sometimes terminate after zero tokens. It appears to be an artifacts
of CFG, since the EOS logit is not particularly large for the conditional
logits; only post-CFG.

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>